First "Dropbox Express" post.
Includes support for long, two-part titles (with an option to
manually provide a short title for the index and other list
contexts), and some styling for code blocks.

Use of `~ yield` in the article layout suppresses Haml white space,
so that code blocks don't get a bunch of extra indentation.